requirements.txt
psutil==6.0.0
watchdog==5.0.3
PyYAML==6.0.2
